requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
google-auth==2.49.2
google-genai==1.14.0
django-tailwind==4.4.2
gunicorn==25.3.0
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
resend==2.27.0
sqlparse==0.5.5
requests==2.32.3
urllib3==2.6.3
whitenoise==6.12.0
youtube-transcript-api==1.2.4
